removed everything from flux console, EM, components, system all live in sandbox
this is how its supposed to be - the engine part of the game is
essentially complete with pygames rendering/event/audio/window/ect...
    systems done. there was never a need to make a class that re-writes
    them from scratch which was essentially replacing the word pygame
    with flux and also essentially reducing functionality and
    transparency. now, i can just make entities/components/systems and
    play around with them. make partical system, make a physics system
    and so on

a more useful approuch is to look at how i can make a scene manager
which is an editor at heart and lets me edit/manipulate levels, save
them and then run them. which is what i will do, but idk how much time
im going to spend on this. why? because even though working on a scene
manager is beneficial and will give me a lot of experience, it doesnt
teach me how to develop those fundamental tools such as
rendering/event/audio/window/ect... systems. but after i do complete
those somewhere else the next step would be a scene manager. so why not
start now and see how it goes.

i appreciate flux for what it was, time to move to better things.