started a rework in the ECS. dont need C's according to Ryan Fleury. just need flat entitiy data.
im not sure how the purple texture thats being draw twice once for each
entity, when being scaled on 1 entity, isnt being scaled on the other.
which is the intended behaviour i just dont know why it works and need
to investigate

i also now load all the textures at once and use them, which is part of
the confusion from the previous line because i load it once and use it
twice

but either way the workflow is setup and seems to work

i really like how i can just open 1 file and add as many variables as i
want and it just works. i dont have to edit mutiple files such as json,
component data for component. its just 1 little file and i can just use
the data.